Rename eome type names and files
