+++ lib/src/Picker/Picker.tsx
@@ -9,8 +9,8 @@ import { BasePickerProps } from '../typings/BasePicker';
 import { MaterialUiPickersDate } from '../typings/date';
 import { DateInputProps } from '../_shared/PureDateInput';
 import { DatePickerView } from '../DatePicker/DatePicker';
-import { WithViewsProps, AnyPickerView } from './WithViewsProps';
 import { useIsLandscape } from '../_shared/hooks/useIsLandscape';
+import { WithViewsProps, AnyPickerView } from './SharedPickerProps';
 import { DIALOG_WIDTH, VIEW_HEIGHT } from '../constants/dimensions';
 import { WrapperVariantContext } from '../wrappers/WrapperVariantContext';
 import { MobileKeyboardInputView } from '../views/MobileKeyboardInputView';
